requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyMuPDF>=1.23.0
arxiv>=2.1.0
pyyaml>=6.0
